+++ test/facemesh/app/src/main/java/com/example/facemesh/MainActivity.java
@@ -280,15 +280,15 @@ public class MainActivity extends AppCompatActivity {
      * @param num - 保留有效数
      * @return
      */
-    private static float getAverage(String type, float[] arr, int num) {
-        float avg = 0f, sum = 0f;
+    private static double getAverage(String type, double[] arr, int num) {
+        double avg = 0f, sum = 0f;
         int len = arr.length;
         for(int i = 0; i < len; i++) {
             sum += arr[i];
         }
         try {
             BigDecimal bigD = new BigDecimal(sum/len);
-            avg = (float) (bigD.setScale(num, BigDecimal.ROUND_HALF_UP).doubleValue());
+            avg = (double) (bigD.setScale(num, BigDecimal.ROUND_HALF_UP).doubleValue());
         } catch (NumberFormatException e) {
 
         }
@@ -302,14 +302,14 @@ public class MainActivity extends AppCompatActivity {
      * @param num - 保留小数点后的有效数位
      * @return
      */
-    private static float getRound(float val, int num) {
-        float ret = 0f;
+    private static double getRound(double val, int num) {
+        double ret = 0f;
         if((num < 1) ||(val == 0f) ||(val == NaN)) {
             return val;
         }
         try {
             BigDecimal bigD = new BigDecimal(val);
-            ret = (float) (bigD.setScale(num, BigDecimal.ROUND_HALF_UP).doubleValue());
+            ret = (double) (bigD.setScale(num, BigDecimal.ROUND_HALF_UP).doubleValue());
         } catch (NumberFormatException e) {
 
         }
@@ -367,9 +367,9 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    private static void getLandmarkMinMax(String key, float[] arr) {
+    private static void getLandmarkMinMax(String key, double[] arr) {
         int cnt = arr.length;
-        float min = 0.0f, max = 0.0f;
+        double min = 0.0f, max = 0.0f;
         for(int i = 0; i < cnt; i++) {
             if((min == 0.0f) && (max == 0.0f)) {
                 min = arr[i];
@@ -386,17 +386,17 @@ public class MainActivity extends AppCompatActivity {
 
     private static final int AVG_CNT = 10;
     private static final int DETECT_TIMES = 2;
-    static float brow_up_arr[] = new float[AVG_CNT];
-    static float brow_width_arr[] = new float[AVG_CNT];
-    static float brow_height_arr[] = new float[AVG_CNT];
-    static float brow_line_arr[] = new float[AVG_CNT];
-    static float brow_mouth_arr[] = new float[AVG_CNT];
-    static float brow_height_mouth_arr[] = new float[AVG_CNT];
-    static float eye_height_arr[] = new float[AVG_CNT];
-    static float eye_width_arr[] = new float[AVG_CNT];
-    static float eye_height_mouth_arr[] = new float[AVG_CNT];
-    static float mouth_width_arr[] = new float[AVG_CNT];
-    static float mouth_height_arr[] = new float[AVG_CNT];
+    static double brow_up_arr[] = new double[AVG_CNT];
+    static double brow_width_arr[] = new double[AVG_CNT];
+    static double brow_height_arr[] = new double[AVG_CNT];
+    static double brow_line_arr[] = new double[AVG_CNT];
+    static double brow_mouth_arr[] = new double[AVG_CNT];
+    static double brow_height_mouth_arr[] = new double[AVG_CNT];
+    static double eye_height_arr[] = new double[AVG_CNT];
+    static double eye_width_arr[] = new double[AVG_CNT];
+    static double eye_height_mouth_arr[] = new double[AVG_CNT];
+    static double mouth_width_arr[] = new double[AVG_CNT];
+    static double mouth_height_arr[] = new double[AVG_CNT];
     static int arr_cnt = 0;
     static int normal_times = 0, suprise_times = 0, sad_times = 0, happy_times = 0, angry_times = 0;
     static int total_log_cnt = 0;
@@ -415,31 +415,31 @@ public class MainActivity extends AppCompatActivity {
         String faceLandmarksStr = "";
         faceLandmarksStr += "\t\tLandmark count: " + landmarkList.getLandmarkCount() + "\n";
         //脸宽
-        float face_width = 0f;
-        float face_height = 0f;
+        double face_width = 0f;
+        double face_height = 0f;
         //眉毛
-        float brow_left_height = 0f;
-        float brow_right_height = 0f;
-        float brow_hight = 0f;
-        float brow_line_left = 0f;
-        float brow_width = 0f;
-        float brow_left_up = 0f;
-        float brow_right_up = 0f;
+        double brow_left_height = 0f;
+        double brow_right_height = 0f;
+        double brow_hight = 0f;
+        double brow_line_left = 0f;
+        double brow_width = 0f;
+        double brow_left_up = 0f;
+        double brow_right_up = 0f;
         //眼睛
-        float eye_left_height = 0f;
-        float eye_left_width = 0f;
-        float eye_right_height = 0f;
-        float eye_right_width = 0f;
-        float eye_height = 0f;
-        float eye_width = 0f;
+        double eye_left_height = 0f;
+        double eye_left_width = 0f;
+        double eye_right_height = 0f;
+        double eye_right_width = 0f;
+        double eye_height = 0f;
+        double eye_width = 0f;
         //嘴巴
-        float mouth_width = 0f;
-        float mouth_height = 0f;
+        double mouth_width = 0f;
+        double mouth_height = 0f;
 
         //眼角嘴角距离
-        float distance_eye_left_mouth = 0f;
-        float distance_eye_right_mouth = 0f;
-        float distance_eye_mouth = 0f;
+        double distance_eye_left_mouth = 0f;
+        double distance_eye_right_mouth = 0f;
+        double distance_eye_mouth = 0f;
 
 //        for (int i = 0; i < landmarkList.getLandmarkCount(); i++) {
 //            faceLandmarksStr  += "\t\tLandmark ["
@@ -482,14 +482,14 @@ public class MainActivity extends AppCompatActivity {
                                 landmarkList.getLandmark(296).getY() - landmarkList.getLandmark(10).getY();
         brow_hight = brow_left_height + brow_right_height;
         //2.2、眉毛高度与识别框高度之比: 眉毛抬高(惊奇、恐惧、悲伤), 眉毛压低(厌恶, 愤怒) - Solution 1(7-1)
-        float brow_hight_rate = (brow_hight/16)/face_width;
-        float brow_width_rate = (brow_width/8)/face_width;
+        double brow_hight_rate = (brow_hight/16)/face_width;
+        double brow_width_rate = (brow_width/8)/face_width;
 //        // 分析挑眉程度和皱眉程度, 左眉拟合曲线(53-52-65-55-70-63-105-66) - 暂时未使用
-//        float line_brow_x[] = new float[3];
+//        double line_brow_x[] = new double[3];
 //        line_brow_x[0] = landmarkList.getLandmark(52).getX();
 //        line_brow_x[1] = landmarkList.getLandmark(70).getX();
 //        line_brow_x[2] = landmarkList.getLandmark(105).getX();
-//        float line_brow_y[] = new float[3];
+//        double line_brow_y[] = new double[3];
 //        line_brow_y[0] = landmarkList.getLandmark(52).getY();
 //        line_brow_y[1] = landmarkList.getLandmark(70).getY();
 //        line_brow_y[2] = landmarkList.getLandmark(105).getY();
@@ -507,7 +507,7 @@ public class MainActivity extends AppCompatActivity {
 
         //2.3、眉毛变化程度: 变弯(高兴、惊奇) - 上扬  - 下拉 - Solution 1(7-2) - 临时关闭(未使用)
         brow_line_left = (landmarkList.getLandmark(105).getY() - landmarkList.getLandmark(52).getY())/(landmarkList.getLandmark(105).getX() - landmarkList.getLandmark(52).getX());
-        float brow_line_rate = brow_line_left;  // + brow_line_right;
+        double brow_line_rate = brow_line_left;  // + brow_line_right;
         brow_left_up = landmarkList.getLandmark(70).getY()-landmarkList.getLandmark(10).getY()/* + landmarkList.getLandmark(66).getY()-landmarkList.getLandmark(10).getY()*/;
         brow_right_up = landmarkList.getLandmark(300).getY()-landmarkList.getLandmark(10).getY()/* + landmarkList.getLandmark(283).getY()-landmarkList.getLandmark(10).getY()*/;
 
@@ -527,7 +527,7 @@ public class MainActivity extends AppCompatActivity {
         mouth_height = landmarkList.getLandmark(17).getY() - landmarkList.getLandmark(0).getY();  // 中心
 
         //4.1、嘴角下拉(厌恶、愤怒、悲伤),    > 1 上扬， < 1 下拉 - Solution 1(7-7)
-        float mouth_line_rate = ((landmarkList.getLandmark(78).getY() + landmarkList.getLandmark(308).getY()))/(landmarkList.getLandmark(14).getY() + landmarkList.getLandmark(0).getY());
+        double mouth_line_rate = ((landmarkList.getLandmark(78).getY() + landmarkList.getLandmark(308).getY()))/(landmarkList.getLandmark(14).getY() + landmarkList.getLandmark(0).getY());
 //        Log.i(TAG, "faceEC: mouth_line_rate = "+mouth_line_rate);
 
         //5、两侧眼角到同侧嘴角距离
@@ -536,20 +536,20 @@ public class MainActivity extends AppCompatActivity {
         distance_eye_mouth = distance_eye_left_mouth + distance_eye_right_mouth;
 
         //6、归一化
-        float MM = 0, NN = 0, PP = 0, QQ = 0;
-        float dis_eye_mouth_rate = (2 * mouth_width)/distance_eye_mouth;             // 嘴角 / 眼角嘴角距离, 高兴(0.85),愤怒/生气(0.7),惊讶(0.6),大哭(0.75)
-        float distance_brow = landmarkList.getLandmark(296).getX() - landmarkList.getLandmark(66).getX();
-        float dis_brow_mouth_rate = mouth_width/distance_brow;                       // 嘴角 / 两眉间距
-        float dis_eye_height_mouth_rate = (1 * mouth_width)/((eye_height)/2);        // 嘴角 / 上下眼睑距离
-        float dis_brow_height_mouth_rate = (2 * mouth_width)/(landmarkList.getLandmark(145).getY() - landmarkList.getLandmark(70).getY());
+        double MM = 0, NN = 0, PP = 0, QQ = 0;
+        double dis_eye_mouth_rate = (2 * mouth_width)/distance_eye_mouth;             // 嘴角 / 眼角嘴角距离, 高兴(0.85),愤怒/生气(0.7),惊讶(0.6),大哭(0.75)
+        double distance_brow = landmarkList.getLandmark(296).getX() - landmarkList.getLandmark(66).getX();
+        double dis_brow_mouth_rate = mouth_width/distance_brow;                       // 嘴角 / 两眉间距
+        double dis_eye_height_mouth_rate = (1 * mouth_width)/((eye_height)/2);        // 嘴角 / 上下眼睑距离
+        double dis_brow_height_mouth_rate = (2 * mouth_width)/(landmarkList.getLandmark(145).getY() - landmarkList.getLandmark(70).getY());
         // 眉毛上扬与识别框宽度之比
-        float brow_up_rate = (brow_left_up + brow_right_up)/(2*face_width);
+        double brow_up_rate = (brow_left_up + brow_right_up)/(2*face_width);
         // 眼睛睁开距离与识别框高度之比
-        float eye_height_rate = eye_height/(2*face_width);
-        float eye_width_rate = eye_width/(2*face_width);
+        double eye_height_rate = eye_height/(2*face_width);
+        double eye_width_rate = eye_width/(2*face_width);
         // 张开嘴巴距离与识别框高度之比
-        float mouth_width_rate = mouth_width/face_width;
-        float mouth_height_rate = mouth_height/face_width;
+        double mouth_width_rate = mouth_width/face_width;
+        double mouth_height_rate = mouth_height/face_width;
 //        Log.i(TAG, "faceEC: 眼角嘴 = "+dis_eye_mouth_rate+", \t眉角嘴 = "+dis_brow_mouth_rate+", \t眼高嘴 = "+dis_eye_height_mouth_rate+", \t眉高嘴 = "+dis_brow_height_mouth_rate);
 
         //7、 求连续多次的平均值
@@ -564,10 +564,10 @@ public class MainActivity extends AppCompatActivity {
         eye_width_arr[arr_cnt] = eye_width_rate;
         mouth_width_arr[arr_cnt] = mouth_width_rate;
         mouth_height_arr[arr_cnt] = mouth_height_rate;
-        float brow_mouth_avg = 0f, brow_height_mouth_avg = 0f;
-        float brow_up_avg = 0f, brow_width_avg = 0f, brow_height_avg = 0f, brow_line_avg = 0f;
-        float eye_height_avg = 0f, eye_width_avg = 0f, eye_height_mouth_avg = 0f;
-        float mouth_width_avg = 0f, mouth_height_avg = 0f;
+        double brow_mouth_avg = 0f, brow_height_mouth_avg = 0f;
+        double brow_up_avg = 0f, brow_width_avg = 0f, brow_height_avg = 0f, brow_line_avg = 0f;
+        double eye_height_avg = 0f, eye_width_avg = 0f, eye_height_mouth_avg = 0f;
+        double mouth_width_avg = 0f, mouth_height_avg = 0f;
         arr_cnt++;
         if(arr_cnt >= AVG_CNT) {
             brow_mouth_avg = getAverage("眉角嘴", brow_mouth_arr, 4);
@@ -585,22 +585,22 @@ public class MainActivity extends AppCompatActivity {
         }
 
         //8、表情算法
-        float brow_height_width_rate = brow_height_avg/brow_width_avg;
-        float eye_width_height_rate = eye_width_avg/eye_height_avg;
-        float mouth_width_height_rate = mouth_width_avg/mouth_height_avg;
+        double brow_height_width_rate = brow_height_avg/brow_width_avg;
+        double eye_width_height_rate = eye_width_avg/eye_height_avg;
+        double mouth_width_height_rate = mouth_width_avg/mouth_height_avg;
 
         if(dis_eye_mouth_rate <= 0.7) {
             MM = dis_eye_mouth_rate * 0;
         } else if((dis_eye_mouth_rate > 0.7) &&(dis_eye_mouth_rate <= 0.75)) {    // 微笑
-            MM = (float)(dis_eye_mouth_rate * 1.38);
+            MM = (double)(dis_eye_mouth_rate * 1.38);
         } else if((dis_eye_mouth_rate > 0.75) &&(dis_eye_mouth_rate <= 0.8)) {
-            MM = (float)(dis_eye_mouth_rate * 2.58);
+            MM = (double)(dis_eye_mouth_rate * 2.58);
         } else if((dis_eye_mouth_rate > 0.8) &&(dis_eye_mouth_rate <= 0.9)) {
-            MM = (float)(dis_eye_mouth_rate * 3.54);
+            MM = (double)(dis_eye_mouth_rate * 3.54);
         } else if((dis_eye_mouth_rate > 0.9) &&(dis_eye_mouth_rate <= 1.0)) {     //大笑
-            MM = (float)(dis_eye_mouth_rate * 4.22);
+            MM = (double)(dis_eye_mouth_rate * 4.22);
         } else if(dis_eye_mouth_rate > 1) {
-            MM = (float)(dis_eye_mouth_rate * 5.0);
+            MM = (double)(dis_eye_mouth_rate * 5.0);
         }
 
         if(brow_height_width_rate <= 0.365f) {
@@ -622,7 +622,7 @@ public class MainActivity extends AppCompatActivity {
         }
 
         //9、判断头部倾斜度
-        float head_line_rate = (landmarkList.getLandmark(362).getY() - landmarkList.getLandmark(133).getY())/(landmarkList.getLandmark(362).getX() - landmarkList.getLandmark(133).getX());
+        double head_line_rate = (landmarkList.getLandmark(362).getY() - landmarkList.getLandmark(133).getY())/(landmarkList.getLandmark(362).getX() - landmarkList.getLandmark(133).getX());
         if(Math.abs(head_line_rate) >= 0.5f) {
             Log.i(TAG, "faceEC: ============头部太偏=============");
             showString = "头部太偏";
